@@ -4,7 +4,9 @@
  */
 
 
-import React, { Component, AppRegistry } from 'react-native';
+import { AppRegistry } from 'react-native';
+import React, { Component } from 'react';
+
 import { Provider } from 'react-redux';
 import Root from './app/components/ios/root.ios';
 import configureStore from './app/utils/configure.store';
